memo all
